% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai2metrics.R
\name{ai2metrics}
\alias{ai2metrics}
\title{ai2metrics: Extraction of aerial image metrics from spectral values linked to 3D point cloud}
\usage{
ai2metrics(pointcloud = NULL, cutoff = 0, min_echo_n = 10, output = NULL)
}
\arguments{
\item{pointcloud}{Specify a point cloud file (text file, delimited with space),
Column structure: plot_cell_id, x, y, z, dz, i, echotype, 
flightline, terraclass, GPS-time, R, G, B, NIR. String. 
Echo type codes: 0 = Only echo, 1 = First of many echo, 
2 = Intermediate echo, 3 = Last of many echo.
Note that the ai2metrics computes the metrics based on the last four columns, 
and they have fixed column names used in the names of output metrics: R, G, B, NIR.}

\item{cutoff}{Cut off all echoes smaller or equal to the given threshold value. Numeric.}

\item{min_echo_n}{Minimum number of echoes to compute metrics. Numeric.}

\item{output}{A path of the output file, including file name as a .txt format. String.}
}
\value{
\code{ai2metrics} prints a .txt file to the user-defined path.
}
\description{
\code{ai2metrics} is used to extract metrics (mean, sd, min , max) from 
the last 4 columns of an input point cloud. The code is intended
for the calculation of spectral metrics fetched from multispectral 
aerial images (R, G, B, NIR). The ai2metrics function uses just first echoes.
}
\examples{
ai2metrics("C:/Temp/lidar_data.txt",    # Point cloud
          2,				                    # Cutoff threshold
          10,                        	# Minimum number of echoes
          "C:/Temp/ai_metrics.txt"     # Output file having ALS metrics
)
}
