% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convlas2txt.R
\name{convlas2txt}
\alias{convlas2txt}
\title{convlas2txt: Convert and combine las/laz files to a format compatible with the als2metrics functions}
\usage{
convlas2txt(
  las_folder = NULL,
  outfile = NULL,
  parse_element = 1,
  rgbn_cols = NULL,
  sub_clip = NULL
)
}
\arguments{
\item{las_folder}{Path of folder that contains las files by plots/units. 
File names must have numeric ID separable by "_". String.}

\item{outfile}{Output file, full name including .txt suffix. String.}

\item{parse_element}{An indicator referring to the ID part or parts of file name. Numeric or a vector of numeric.}

\item{rgbn_cols}{An indicator vector referring to the columns added 
to the places of R, G, B, NIR in the output file (see the format in the Return section). 
 Numeric vector or NULL}

\item{sub_clip}{A list including arguments for sub-clipping, e.g. circular plot. Note that no overlap checks are carried out.
The code assumes that there exists a polygon for sub-clipping for each plot/unit file. The linkage between las/laz files
and polygons are made based on IDs (id_col). The sub_id_col argument is used if several polygons per txt file exist.
 An example argument: list(path_pols = "C:/Temp/circ_plots.gpkg", id_col = 1, sub_id_col = 2, crs = 3067).
 Note: id_col and sub_id_col strings will be concatenated using "_".}
}
\value{
\code{convlas2txt} outputs a .txt file to the user-defined path. 
The output file follows the format: id, x, y, z, dz, i, echotype, 
flightline, terraclass, numofret, retnum, R, G, B, NIR.
Returns the outfile path (String) that can be passed to the other als2metrics functions.
}
\description{
\code{convlas2txt} is used to convert multiple las/laz files into the txt format used in the als2metrics and ai2metrics functions.
}
\examples{
convertlas2txt(las_folder = "C:/Temp/las_plots", # files: 1_plot.las, 2_plot.las... 
              outfile = "C:/Temp/las_plots/plotpoints.txt", 
              parse_element = 1,                # 2_plot.las, parse_element value 1 collects ID = 2 from the name of the las file.
              rgbn_cols = c(16, 17, 18, 19))    # For example. If NULL, function returns values -1 for the last four columns.
)
With a sub-clipping
convertlas2txt(las_folder = "C:/Temp/las_plots", # files: 1_plot.las, 2_plot.las... 
              outfile = "C:/Temp/las_plots/plotpoints.txt", 
              parse_element = 1,                # 2_plot.las, parse_element value 1 collects ID = 2 from the name of the las file.
              rgbn_cols = c(16, 17, 18, 19),    # For example. If NULL, function returns values -1 for the last four columns.
              sub_clip = list(path_pols = "C:/Temp/circ_plots.gpkg", id_col = 1, crs = 3067)) # id in the first column
)
convertlas2txt(las_folder = "C:/Temp/las_plots", # files: 1_plot.las, 2_plot.las... 
              outfile = "C:/Temp/las_plots/plotpoints.txt", 
              parse_element = c(1,2),                # plot_3_plot.las, parse_element specified as c(1, 2) returns "plot_3" for the processing
              sub_clip = list(path_pols = "C:/Temp/circ_plots.gpkg", id_col = 1, sub_id_col = 3, crs = 3067)) # id in the first column, one pol per tile
              # sub_id_col is used if there are several clips per plot
}
