% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcliptxt.R
\name{subcliptxt}
\alias{subcliptxt}
\title{subcliptxt: Convert and combine txt files to a format compatible with the als2metrics functions}
\usage{
subcliptxt(
  txt_folder = NULL,
  outfile = NULL,
  parse_element = 1,
  sub_clip = NULL
)
}
\arguments{
\item{txt_folder}{Path of folder that contains las files by plots/units. 
File names must have numeric ID separable by "_". String.}

\item{outfile}{Output file, full name including .txt suffix. String.}

\item{parse_element}{An indicator referring to the ID part of file name. Numeric.}

\item{sub_clip}{A list including arguments for sub-clipping, e.g. circular plot. Note that no overlap checks are carried out.
The code assumes that there exists a polygon for sub-clipping for each plot/unit file. The linkage between las/laz files
and polygons are made based on IDs (id_col). The sub_id_col argument is used if several polygons per txt file exist.
 An example argument: list(path_pols = "C:/Temp/circ_plots.gpkg", id_col = 1, sub_id_col = 2, crs = 3067).
Note: id_col and sub_id_col strings will be concatenated using "_".}
}
\value{
\code{subcli2txt} outputs a .txt file to the user-defined path. 
The output file follows the format: id, x, y, z, dz, i, echotype, 
flightline, terraclass, numofret, retnum, R, G, B, NIR.
Returns the outfile path (String) that can be passed to the other als2metrics functions.
}
\description{
\code{convlas2txt} is used to convert multiple txt files (tiles) into the txt format used in the als2metrics and ai2metrics functions.
The function also allows subclips based on polygons.
}
\examples{
subcliptxt(txt_folder = "C:/Temp/las_plots", # files: 1_plot.txt, 2_plot.txt... 
              outfile = "C:/Temp/las_plots/plotpoints.txt", 
              parse_element = 1)                # 2_plot.las, parse_element value 1 collects ID = 2 from the name of the las file.
)
With a sub-clipping
subcliptxt(txt_folder = "C:/Temp/las_plots", # files: 1_plot.las, 2_plot.las... 
              outfile = "C:/Temp/las_plots/plotpoints.txt", 
              parse_element = 1,                # 2_plot.las, parse_element value 1 collects ID = 2 from the name of the las file.
              sub_clip = list(path_pols = "C:/Temp/circ_plots.gpkg", id_col = 1, crs = 3067)) # id in the first column, one pol per tile
              # sub_id_col is used if there are several clips per plot
)
}
