% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/als2metrics.R
\name{als2metrics}
\alias{als2metrics}
\title{als2metrics: Extract metrics from an airborne LiDAR point cloud}
\usage{
als2metrics(
  ALSFILE = NULL,
  FIRST = NULL,
  LAST = NULL,
  INTERMEDIATE = NULL,
  ALL_ECHOES = NULL,
  PROP_MEAN_ETYP = NULL,
  BASIC_STATISTICS = NULL,
  PERCENTILE_SCALE = NULL,
  DENSITIES = NULL,
  DENSITIES_FIXED_TRESHOLD = NULL,
  INTENSITY_STATISTICS = NULL,
  INTENSITY_PERCENTILES = NULL,
  CUTOFF = NULL,
  MIN_ECHO_N = NULL,
  output = NULL
)
}
\arguments{
\item{ALSFILE}{Specify an ALS file (text file), Format: plot_cell_id; x; y; z; dz, i; 
echotype; flightline; terraclass; GPS-time (delimeter: space). 
First seven columns must be in the abovementioned order, additional columns are optional.
Echo type codes: 0 = Only echo, 1 = First of many echo, 
2 = Intermediate echo, 3 = Last of many echo. String.}

\item{FIRST}{Output metrics computed based on  FIRST echoes (first + only)
(rows). It is recommended to use row and column names. Logical.}

\item{LAST}{Output metrics computed based on  LAST echoes (last + only). Logical.}

\item{INTERMEDIATE}{Output metrics computed based on INTERMEDIATE echoes (intermediate). Logical.}

\item{ALL_ECHOES}{Output metrics computed based on ALL echoes (first, last, only and intermediate). Logical.}

\item{PROP_MEAN_ETYP}{Calculate mean and standard deviation of heights and the proportion of echoes categories. Logical.}

\item{BASIC_STATISTICS}{Calculates mean, std, med, min, max, skew, kurt. Logical.}

\item{PERCENTILE_SCALE}{A vector of percentiles, e.g using seq(...) function. Percentiles are calculated using quantile() -function (using default type=7)}

\item{DENSITIES}{Calculates densities, i.e. echo proportion under or equal to the determined height value. Logical.}

\item{DENSITIES_FIXED_TRESHOLD}{The height values fixed height densities are computed. Vector.}

\item{INTENSITY_STATISTICS}{Calculates mean_int, std_int, med_int, min_int, max_int, skew_int, kurt_int}

\item{INTENSITY_PERCENTILES}{Calculate perecentiles in the same manner as for dZ values ('Compute percentiles' must be set to TRUE). Logical.}

\item{CUTOFF}{Cut off all echoes smaller or equal to the given threshold value. Numeric.}

\item{MIN_ECHO_N}{Minimum number of echoes to compute metrics. Numeric.}

\item{output}{A path of the output file, including file name as a .txt format. String.}
}
\value{
\code{als2metrics} Prints a .txt file to the user-defined path.
}
\description{
\code{als2metrics} is used to extract metrics from an airborne LiDAR point cloud
}
\examples{
als2metrics(ALSFILE = "C:/Temp/lidar_data.txt",            # A Path of an airborne LiDAR data file
           FIRST = TRUE,                        			 # Compute first echo metrics
           LAST = TRUE,                        			 # Compute last echo metrics
           INTERMEDIATE = TRUE,                           # Compute intermediate echo metrics
           ALL_ECHOES = TRUE,                             # Compute all echo metrics
           PROP_MEAN_ETYP = TRUE,                         # Compute proportions of echo categories
           BASIC_STATISTICS = TRUE,                       # Compute basic statistics
           PERCENTILE_SCALE = seq(0.05, 0.95, 0.05),      # Vector of percentiles
           DENSITIES = TRUE,                              # Compute densities
           DENSITIES_FIXED_TRESHOLD = c(0.5, 2, 5, 10, 15, 20),    # Vector of heights in fixed height densities
           INTENSITY_STATISTICS = TRUE,                            # Compute intensity statistics
           INTENSITY_PERCENTILES = TRUE,                           # Compute intensity percentiles
           CUTOFF = 0.0,                                  # Cutoff threshold
           MIN_ECHO_N = 10,                               # Minimum number of echoes
           output = "lidar_metrics.txt"                   # Output file 
)
}
